refactor: common types
