+++ src/models/albums/AlbumsResolver.ts
@@ -1,7 +1,7 @@
 import { UserInputError } from 'apollo-server-core';
 import { Resolvers } from '../../types/resolvers';
 import { Album, CreateAlbumArgs, UpdateAlbumArgs } from '../../types/albums';
-import { QueryParams } from '../../types/queryParams';
+import { QueryParams } from '../../types/common';
 import InputError from '../../const/errors';
 import filterByExistance from '../../utils/firlterByExistanse';
