+++ src/models/bands/bandsService.ts
@@ -1,9 +1,7 @@
 import { RESTDataSource, RequestOptions } from 'apollo-datasource-rest';
 import { API, BandsAPIEndpoint } from '../../const/api';
 import { Band, CreateBandArgs, UpdateBandArgs } from '../../types/bands';
-import { DeleteResponse } from '../../types/deleteResponse';
-import { List } from '../../types/list';
-import { QueryParams } from '../../types/queryParams';
+import { DeleteResponse, List, QueryParams } from '../../types/common';
 import getQueryParams from '../../utils/getQueryParams';
 
 export interface BandsAPIDataSource {
