+++ src/models/genres/GenresReolver.ts
@@ -1,7 +1,7 @@
 import { UserInputError } from 'apollo-server-core';
 import { Resolvers } from '../../types/resolvers';
 import { CreateGenreArgs, Genre, UpdateGenreArgs } from '../../types/genres';
-import { QueryParams } from '../../types/queryParams';
+import { QueryParams } from '../../types/common';
 import InputError from '../../const/errors';
 
 const genresResolver: Resolvers = {
