--- src/types/deleteResponse.ts
@@ -1,4 +1,0 @@
-export type DeleteResponse = {
-  acknowledged: boolean
-  deletedCound: number;
-};
