--- src/types/list.ts
@@ -1,6 +1,0 @@
-export type List<T> = {
-  items: T[];
-  limit: number;
-  offset: number
-  total: number;
-};
