--- src/types/queryParams.ts
@@ -1,4 +1,0 @@
-export type QueryParams = {
-  limit?: number;
-  offset?: number;
-};
